+++ new-merchant-quote-service-api/src/main/java/nz/co/smartpay/validator/QuoteRequestValidator.java
@@ -2,10 +2,12 @@ package nz.co.smartpay.validator;
 
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class QuoteRequestValidator {
 
-    public void validate(String industry, Long transactionCount, Long transactionVolume) {
+    public void validate(String industry, Long transactionCount, BigDecimal transactionVolume) {
         // todo - validate
     }
 }
